adding more options

+++ lib/heroku_clonedb.rb
@@ -22,41 +22,88 @@ module Heroku::Command
     def index
       
       is_root?
-      
-      display("Warning: Data in the app '#{app}' will be overwritten and will not be recoverable.")
-      
-      exit unless confirm
 
       opts             = parse_database_yml
       database_options = load_database_options(opts)
-
-      dump_name = "#{opts[:database]}.dump"
-
-      display "===== Capture Backup from #{app}...", false
-       
-      run "heroku pgbackups:capture --expire --app #{app}"
-
-      display "===== Downloading Data Dump...", false
-
-      run "curl -o tmp/#{dump_name} `heroku pgbackups:url --app #{app}`"
-
-      display "===== Deleting database #{opts[:database]}...", false
+      dump_name        = "#{opts[:database]}.dump"
       
-      run "dropdb #{opts[:database]} #{database_options}"
+      from_app = extract_option("--from") || app
+      from_url = extract_option("--from-url")
+      to_app   = extract_option("--to")
+      
+      if from_app.nil? && from_url.nil? && to_app.nil?
+        raise(CommandFailed, help)
+      end
 
-      display "===== Creating database #{opts[:database]}...", false
-      
-      run "createdb #{opts[:database]} #{database_options}"
+      if from_app && to_app
+        has_pgbackups_addon?(from_app)
+        has_pgbackups_addon?(to_app)
 
-      display "===== Restoring database #{opts[:database]} with #{dump_name}...", false
+        display "===== Capture Backup from #{from_app}...", false
+
+        run "heroku pgbackups:capture --expire --app #{from_app}"
+
+        display "===== Transfering data from #{from_app} to #{to_app}...", false
+
+        system "heroku pgbackups:restore DATABASE `heroku pgbackups:url --app #{from_app}` --app #{to_app}"
+
+      elsif from_app
+
+        has_pgbackups_addon?(from_app)
+
+        display("Warning: Data in the local db '#{opts[:database]}' will be overwritten and will not be recoverable with data #{from_app} app.")
+
+        exit unless confirm
+
+        display "===== Capture Backup from #{from_app} app...", false
+
+        run "heroku pgbackups:capture --expire --app #{from_app}"
+
+        display "===== Downloading Data Dump...", false
+
+        run "curl -o tmp/#{dump_name} `heroku pgbackups:url --app #{from_app}`"
+
+        display "===== Restoring database #{opts[:database]} with #{dump_name}...", false
       
-      shell "pg_restore -d #{opts[:database]} tmp/#{dump_name} -O --verbose --clean --no-acl --no-owner > /dev/null 2>&1"
-      shell "rm -f tmp/#{dump_name}"
+        shell "pg_restore -d #{opts[:database]} tmp/#{dump_name} -O --verbose --clean --no-acl --no-owner > /dev/null 2>&1"
+        shell "rm -f tmp/#{dump_name}"
       
-      display "[OK]"
-            
+        display "[OK]"
+
+      elsif to_app && from_url
+        has_pgbackups_addon?(to_app)
+
+        display "===== Loading data to #{to_app} app from url...", false
+      
+        system "heroku pgbackups:restore DATABASE '#{from_url}' --app #{to_app}"
+
+      end
     end
 
+    def dump
+      is_root?
+
+      from_app = extract_option("--from")
+
+      raise(CommandFailed, "No --from app specified.\nYou need define --from <app name>") unless from_app
+
+      has_pgbackups_addon?(from_app)
+
+      #define vars
+      opts             = parse_database_yml
+      dump_name        = "#{opts[:database]}-#{Date.today.to_s}.dump"
+
+      shell "heroku pgbackups:capture --expire --app #{from_app}"
+
+      if dir = extract_option("--dir")
+        run "curl -o #{dir}/#{dump_name} `heroku pgbackups:url --app #{from_app}`"
+        display "The dump is #{dir}/#{dump_name}", false
+      else
+        run "curl -o #{dump_name} `heroku pgbackups:url --app #{from_app}`"
+        display "You can found the dump in #{dump_name}", false
+      end
+      puts ""
+    end
 
     private
 
@@ -67,6 +114,15 @@ module Heroku::Command
       end
     end
     
+    def has_pgbackups_addon?(app_option = nil)
+      app_option = app_option ? app_option : app
+      pgbackups_addon = "pgbackups"
+      unless heroku.installed_addons(app_option).select{|addon| addon["name"].match(/pgbackups/)}.first
+        display "Adding #{pgbackups_addon} to #{app_option}... ", true
+        heroku.install_addon(app_option, pgbackups_addon, {})
+      end
+    end
+
     def run(cmd)
       shell cmd
       if $?.exitstatus == 0
@@ -100,6 +156,44 @@ module Heroku::Command
       puts "Error parsing database.yml: #{ex.message}"
       puts ex.backtrace
       ""
+    end
+
+    def extract_option(options, default=true)
+      values = options.is_a?(Array) ? options : [options]
+      return unless opt_index = args.select { |a| values.include? a }.first
+      opt_position = args.index(opt_index) + 1
+      if args.size > opt_position && opt_value = args[opt_position]
+        if opt_value.include?('--')
+          opt_value = nil
+        else
+          args.delete_at(opt_position)
+        end
+      end
+      opt_value ||= default
+      args.delete(opt_index)
+      block_given? ? yield(opt_value) : opt_value
+    end
+
+    def help
+help = <<EOF
+For load a data dump from heroku app to localhost:
+
+    $ heroku clonedb --from app_name
+
+For transfer data from one app to another:
+
+    $ heroku clonedb --from app_name1 --to app_name2
+
+For transfer data from url dump to heroku app:
+
+    $ heroku clonedb --from-url 'http://s3.amazonaws.com/.....mydb.dump?authparameters' --to app_name
+
+For get a data dump from heroku app:
+
+    $ heroku clonedb:dump --from app_name [--dir <dir_path>]
+
+EOF
+      help
     end
 
   end
